+++ App/Controllers/FastController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Controllers;
 
 date_default_timezone_set('Etc/GMT+0');
