+++ App/Controllers/MainController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Controllers;
 
 
 use App\Console\Input;
